+++ ORM/LiveJobsGridSource.php
@@ -3,15 +3,17 @@
 namespace Dtc\QueueBundle\ORM;
 
 use Doctrine\ORM\EntityManager;
+use Dtc\GridBundle\Grid\Column\GridColumn;
 use Dtc\GridBundle\Grid\Source\EntityGridSource;
 
 class LiveJobsGridSource extends EntityGridSource
 {
     protected $jobManager;
+    protected $running = false;
 
     public function getId()
     {
-        return 'dtc_queue.grid_source.live_jobs.orm';
+        return 'dtc_queue.grid_source.jobs_'.($this->isRunning() ? 'running' : 'waiting').'.orm';
     }
 
     public function __construct(JobManager $jobManager)
@@ -32,14 +34,58 @@ class LiveJobsGridSource extends EntityGridSource
         return parent::getColumns();
     }
 
+    public function setRunning($flag)
+    {
+        $this->running = $flag;
+    }
+
+    public function isRunning()
+    {
+        return $this->running;
+    }
+
+    protected function getRunningQueryBuilder()
+    {
+        /** @var EntityManager $entityManager */
+        $entityManager = $this->jobManager->getObjectManager();
+        $queryBuilder = $entityManager->createQueryBuilder()->add('select', 'j');
+        $queryBuilder->from($this->jobManager->getJobClass(), 'j');
+        $queryBuilder->orderBy('j.startedAt', 'DESC');
+        $queryBuilder->setFirstResult($this->offset)
+                     ->setMaxResults($this->limit);
+
+        return $queryBuilder;
+    }
+
     protected function getQueryBuilder()
     {
+        if ($this->isRunning()) {
+            return $this->getRunningQueryBuilder();
+        }
+
         $queryBuilder = $this->jobManager->getJobQueryBuilder();
         $queryBuilder->add('select', 'j');
         $queryBuilder->setFirstResult($this->offset)
                      ->setMaxResults($this->limit);
 
         return $queryBuilder;
+    }
+
+    public function autoDiscoverColumns()
+    {
+        parent::autoDiscoverColumns();
+        if ($this->columns) {
+            foreach ($this->columns as $column) {
+                if ($column instanceof GridColumn) {
+                    $column->setOption('sortable', false);
+                }
+            }
+        }
+    }
+
+    public function getDefaultSort()
+    {
+        return null;
     }
 
     public function getCount()