+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-### 3.0.3
+### 3.1.0
+   * New Archive button on the Live Jobs administration page to archive all jobs at once
+   * Fix a bug when running doctrine orm and odm simulatenously
+### 3.0.4
    * Refactoring out the TrendsController from the QueueController, ran php-cs-fixer
 ### 3.0.2
    * Updates to trends page, GridBundle version, WorkerCompilerTest